Refactoring CameraRGBD class (now including OpenNI, OpenNI2, OpenNI from OpenCV and Freenect) Added tool to test RGB-D camera: rtabmap-rgbd_camera Fixed Freenect corrupted depth image (after some time)
git-svn-id: http://rtabmap.googlecode.com/svn/trunk/rtabmap@1366 f169173b-cf89-36c8-b27e-44dbe73f0c83

+++ corelib/src/Features2d.cpp
@@ -54,7 +54,7 @@ void filterKeypointsByDepth(
 		float depthConstant,
 		float maxDepth)
 {
-	if(!depth.empty() && depthConstant > 0.0f && maxDepth > 0.0f && (descriptors.empty() || descriptors.rows == keypoints.size()))
+	if(!depth.empty() && depthConstant > 0.0f && maxDepth > 0.0f && (descriptors.empty() || descriptors.rows == (int)keypoints.size()))
 	{
 		std::vector<cv::KeyPoint> output(keypoints.size());
 		std::vector<int> indexes(keypoints.size(), 0);
@@ -71,7 +71,7 @@ void filterKeypointsByDepth(
 		output.resize(oi);
 		keypoints = output;
 
-		if(!descriptors.empty() && keypoints.size() != descriptors.rows)
+		if(!descriptors.empty() && (int)keypoints.size() != descriptors.rows)
 		{
 			if(keypoints.size() == 0)
 			{
